Refactor CLI: init & generate_configs commands
